separate tests by type (integration, unit, e2e)
