+++ .github/workflows/ci.yml
@@ -76,7 +76,6 @@ jobs:
     env:
       LOG_LEVEL: DEBUG
       STREAM_LOG: True
-      QISKIT_TESTS: skip_online
       QISKIT_IN_PARALLEL: True
     steps:
       - uses: actions/checkout@v2
