+++ test/types/e2e/__init__.py
@@ -1,0 +1,18 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2022.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""E2E tests.
+
+Executed against an external system configured via the (token, instance, url) tuple. Basic coverage of
+most important user-facing happy paths. Test suite runs faster than integration but slower than unit
+tests and is stable.
+"""
