+++ test/types/e2e/test_tutorials.py
@@ -12,19 +12,15 @@
 
 """Tests for the tutorials, copied from ``qiskit-iqx-tutorials``."""
 
-from unittest import skipIf
-import os
 import glob
+import os
 import warnings
-
-from qiskit.test.decorators import TEST_OPTIONS
 
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor
 
 from qiskit_ibm_runtime.utils.utils import to_python_identifier
-
-from .ibm_test_case import IBMTestCase
+from ...ibm_test_case import IBMTestCase
 
 TUTORIAL_PATH = "docs/tutorials/**/*.ipynb"
 
@@ -51,7 +47,6 @@ class TutorialsTestCaseMeta(type):
         return type.__new__(mcs, name, bases, dict_)
 
 
-@skipIf(not TEST_OPTIONS["run_slow"], "Skipping slow tests.")
 class TestTutorials(IBMTestCase, metaclass=TutorialsTestCaseMeta):
     """Tests for tutorials."""
