+++ test/types/integration/test_interim_results.py
@@ -16,10 +16,10 @@ import time
 
 from qiskit.providers.jobstatus import JobStatus
 
-from .ibm_test_case import IBMIntegrationJobTestCase
-from .utils.decorators import run_integration_test
-from .utils.utils import cancel_job_safe, wait_for_status
-from .mock.proxy_server import MockProxyServer, use_proxies
+from ..unit.mock.proxy_server import MockProxyServer, use_proxies
+from ...ibm_test_case import IBMIntegrationJobTestCase
+from ...utils.decorators import run_integration_test
+from ...utils.utils import cancel_job_safe, wait_for_status
 
 
 class TestIntegrationInterimResults(IBMIntegrationJobTestCase):
