+++ test/types/integration/test_program.py
@@ -12,19 +12,18 @@
 
 """Tests for runtime service."""
 
-import unittest
 import os
 import tempfile
+import unittest
 
 from qiskit_ibm_runtime.exceptions import IBMNotAuthorizedError
-from qiskit_ibm_runtime.runtime_program import RuntimeProgram
 from qiskit_ibm_runtime.exceptions import (
     RuntimeProgramNotFound,
 )
-
-from .ibm_test_case import IBMIntegrationTestCase
-from .utils.decorators import run_integration_test
-from .utils.templates import RUNTIME_PROGRAM, PROGRAM_PREFIX
+from qiskit_ibm_runtime.runtime_program import RuntimeProgram
+from ...ibm_test_case import IBMIntegrationTestCase
+from ...utils.decorators import run_integration_test
+from ...utils.templates import RUNTIME_PROGRAM, PROGRAM_PREFIX
 
 
 class TestIntegrationProgram(IBMIntegrationTestCase):
