+++ test/types/unit/test_backend_retrieval.py
@@ -12,13 +12,13 @@
 
 """Backends Filtering Test."""
 
-from qiskit.test.mock.backends import FakeLima
 from qiskit.providers.exceptions import QiskitBackendNotFoundError
+from qiskit.test.mock.backends import FakeLima
 
-from .ibm_test_case import IBMTestCase
 from .mock.fake_account_client import BaseFakeAccountClient
 from .mock.fake_runtime_service import FakeRuntimeService
-from .utils.decorators import run_legacy_and_cloud_fake
+from ...ibm_test_case import IBMTestCase
+from ...utils.decorators import run_legacy_and_cloud_fake
 
 
 class TestBackendFilters(IBMTestCase):
