+++ test/types/unit/test_client_parameters.py
@@ -19,7 +19,7 @@ from qiskit_ibm_runtime.proxies import ProxyConfiguration
 from qiskit_ibm_runtime.api.client_parameters import ClientParameters
 from qiskit_ibm_runtime.api.auth import CloudAuth, LegacyAuth
 
-from .ibm_test_case import IBMTestCase
+from ...ibm_test_case import IBMTestCase
 
 
 class TestClientParameters(IBMTestCase):
