+++ test/types/unit/test_job_retrieval.py
@@ -13,11 +13,10 @@
 """Tests for runtime job retrieval."""
 
 from qiskit_ibm_runtime.exceptions import IBMInputValueError
-
-from .ibm_test_case import IBMTestCase
 from .mock.fake_runtime_service import FakeRuntimeService
-from .utils.program import run_program, upload_program
-from .utils.decorators import run_legacy_and_cloud_fake
+from ...ibm_test_case import IBMTestCase
+from ...utils.decorators import run_legacy_and_cloud_fake
+from ...utils.program import run_program, upload_program
 
 
 class TestRetrieveJobs(IBMTestCase):
