+++ test/types/unit/test_logger.py
@@ -12,15 +12,14 @@
 
 """Tests related to logger setup via ``setup_logger()``."""
 
-import os
 import logging
+import os
 from tempfile import NamedTemporaryFile
 from unittest import skipIf, mock
 
 from qiskit_ibm_runtime import QISKIT_IBM_RUNTIME_LOG_LEVEL, QISKIT_IBM_RUNTIME_LOG_FILE
 from qiskit_ibm_runtime.utils.utils import setup_logger
-
-from .ibm_test_case import IBMTestCase
+from ...ibm_test_case import IBMTestCase
 
 
 class TestLogger(IBMTestCase):
