+++ test/types/unit/test_runtime_ws.py
@@ -17,11 +17,9 @@ import time
 from qiskit.test.mock.fake_qasm_simulator import FakeQasmSimulator
 
 from qiskit_ibm_runtime import RuntimeJob
-from qiskit_ibm_runtime.exceptions import RuntimeInvalidStateError
 from qiskit_ibm_runtime.api.client_parameters import ClientParameters
-
-from .ibm_test_case import IBMTestCase
-from .mock.ws_server import MockWsServer
+from qiskit_ibm_runtime.exceptions import RuntimeInvalidStateError
+from .mock.fake_runtime_client import BaseFakeRuntimeClient
 from .mock.ws_handler import (
     websocket_handler,
     JOB_ID_PROGRESS_DONE,
@@ -30,7 +28,8 @@ from .mock.ws_handler import (
     JOB_ID_RETRY_FAILURE,
     JOB_PROGRESS_RESULT_COUNT,
 )
-from .mock.fake_runtime_client import BaseFakeRuntimeClient
+from .mock.ws_server import MockWsServer
+from ...ibm_test_case import IBMTestCase
 
 
 class TestRuntimeWebsocketClient(IBMTestCase):
